requirements.txt
aiofiles==22.1.0
aiosqlite==0.18.0
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
comm==0.1.2
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
fonttools==4.25.0
idna==3.4
ipykernel==6.19.2
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.6
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.1.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyter-ydoc==0.2.4
jupyterlab==3.6.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.22.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==0.8.4
munkres==1.1.4
nbclassic==0.5.5
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.3
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.16.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.1.2
platformdirs==2.5.2
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.36
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.2
QtPy==2.2.0
requests==2.29.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.0
setuptools==67.8.0
shapely==2.0.1
sip==6.6.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.4
stack-data==0.2.0
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
typing_extensions==4.6.3
urllib3==1.26.16
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.38.4
widgetsnbextension==4.0.5
y-py==0.5.9
ypy-websocket==0.8.2
